🐛 fix: address code review issues in layer controls
- Add .layer-controls-btn to focus-visible CSS rules for keyboard a11y
- Reset layerVisibility on campaign change in MainEditor
- Move LayerVisibility type to src/types/LayerVisibility.ts per convention
