+++ src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './Campaign';
+export * from './LayerVisibility';
 export * from './MapExport';
 export * from './Markers';
 export * from './Quest';
